requirements.txt
Flask===3.0.0
Flask-SQLAlchemy===3.1.1
Flask-Migrate===4.0.5
Flask-Script===2.0.6
PyMySQL===1.1.0
cryptography===41.0.7
python-dotenv===1.0.0
dynaconf===3.2.4
flask-simplelogin===0.1.2
sqlalchemy_serializer===1.4.1
Flask-RESTful===0.3.10
flask_debugtoolbar===0.14.1
Flask-Cors===4.0.0

